Implement Font::DrawTextString/DrawScaledTextString; fix cursor int type
Reconstruct Nu3D::Font::DrawTextString (0x004B4CD0) and
DrawScaledTextString (0x004B45A0). Both iterate the input string,
advancing g_textCursorX by each glyph's drawn width (via the
Draw*Glyph helpers, selected by a Compute*CharClip visibility test),
tracking the widest line and returning it. Tab rounds the cursor up to
the next g_textTabWidth boundary; '\n' advances Y by font->fontHeight
(unscaled) or ftol(g_scaledFontHeight) (scaled) and resets X to the
paragraph start; '\r' advances Y by ftol(g_scaledFontHeight) and
resets X to g_textClipX1. The \n/\r line-reset tail is shared via a
goto label, matching retail's fallthrough/jump block.

Corrects a pre-existing type error: g_textCursorX/g_textCursorY are
int32_t, not float. Evidence: DrawTextString does raw integer
ADD/IDIV on the cursor, the draw primitives FILD the cursor (int->float)
for vertex setup, and the ProcessMiscEventsEx caller runs g_scaledFontAscent
through __ftol before passing it to SetTextCursor. SetTextCursor's
signature becomes (int32_t,int32_t) and DrawTextStringFunc now returns
int32_t (both functions return the text width in EAX, mirroring the
Calculate*TextSize family).

Adds the g_textTabWidth global (0x00508664) and forward-declares the
six per-glyph helpers (Compute*CharClip/Draw*Glyph/DrawClipped*Glyph)
as STUBs so the drivers can reference them.

Comparison: DrawTextString 29.79%, DrawScaledTextString 36.76% --
remaining diff is pure register allocation (sub esp vs push ecx frame,
char spill to stack, cursorX reloaded from memory) plus the else-branch
max(x,x) optimizer fold, identical semantics. SetTextCursor stays
100%; ProcessMiscEventsEx improves 73.33% -> 78.01% from the __ftol fix.

+++ src/Toy2/Toy2.cpp
@@ -1186,7 +1186,7 @@ namespace Toy2
 	// FUNCTION: TOY2 0x004909E0
 	void ProcessMiscEventsEx()
 	{
-		Nu3D::Font::SetTextCursor(0, Nu3D::g_scaledFontAscent);
+		Nu3D::Font::SetTextCursor(0, (int32_t)Nu3D::g_scaledFontAscent);
 		DevDraw::g_vertexCount = 0;
 
 		D3DApp::g_windowData.wndIsExiting = g_wndIsExitingUnused;
